removed the old basin setup, cleaning up domain extent
