+++ .codeclimate.yml
@@ -1,0 +1,5 @@
+version: "2"
+checks:
+    method-complexity:
+        config:
+            threshold: 8
